compile errors for uac2_speaker

+++ examples/device/uac2_speaker_fb/src/common_types.h
@@ -43,9 +43,9 @@ typedef struct
     uint8_t alt_settings;
     int8_t mute[CFG_TUD_AUDIO_FUNC_1_N_CHANNELS_RX + 1];
     int16_t volume[CFG_TUD_AUDIO_FUNC_1_N_CHANNELS_RX + 1];
-    uint16_t fifo_size;
-    uint16_t fifo_count;
-    uint16_t fifo_count_avg;
+    uint32_t fifo_size;
+    uint32_t fifo_count;
+    uint32_t fifo_count_avg;
   } audio_debug_info_t;
 #endif
